feat: onboard 18 new Bailian models and fix usage-stats int4 overflow
New models (verified upstream + smoke-billed end-to-end):
- chat: qwen3.8-flash, glm-5.3, deepseek-v4.1-flash (idle-price
  convention), kimi-k2.7-code, kimi/kimi-k2.7-code-highspeed,
  kimi-k2-thinking, MiniMax/MiniMax-M2.7 (slash id, upstream
  rejects the plain form)
- image: wan2.7-image, wan2.7-image-pro
- video: wan2.7-t2v/i2v/r2v/videoedit, wan3.0-video(-prime),
  happyhorse-1.1-t2v/i2v/r2v with async-billing unit-price
  branches (new prefixes must precede nothing; kept after
  wan2.6- exact matches)

Bug fix: SUM(total_tokens)::int overflows int4 at current volume
(3.55B tokens). Cast to ::float8 in getDaily/getByModel/
getUsageSummary/getModelDistribution so dashboards stop 500ing.

MiniMax-H3 dropped: it is an omni video model that dashscope does
not expose at all.

+++ backend/src/data/dashboard.ts
@@ -225,7 +225,7 @@ export async function getModelDistribution() {
     `SELECT
        model,
        COUNT(*)::int as requests,
-       COALESCE(SUM(total_tokens), 0)::int as tokens,
+       COALESCE(SUM(total_tokens), 0)::float8 as tokens,
        ROUND(COALESCE(SUM(cost), 0)::numeric, 2)::float as cost
      FROM usage_logs
      WHERE created_at >= NOW() - INTERVAL '7 days'
